Represent gas as a long integer.
- There is therectially no way to consume more than Long.MAX_VALUE gas
- Some intemediary gas computation is still in BigInteger, to avoid arithmetic overflow
- There are only two sources of user-defined gas: 1) in the transaction; 2) via CALL opcode. Sanity check will be required there.

+++ src/main/java/org/ethereum/vm/client/Transaction.java
@@ -70,13 +70,11 @@ public interface Transaction {
     byte[] getData();
 
     /**
-     * Returns the gas limit. Because an invalid internal transaction will set the
-     * gas limit larger than {@link Long#MAX_VALUE}, this value is represented in
-     * BigInt.
+     * Returns the gas limit.
      *
      * @return the specified gas for this transaction.
      */
-    BigInteger getGas();
+    long getGas();
 
     /**
      * Returns the gas price.
